Get Member, workspace details, delineate modules
Already created separate scapi.ari and scdata.ari files. This commit
starts a new sc.ari which loads both and is meant to be where more of
the polished interactive experience is developed. The scapi.ari file
should remain sufficient for anyone looking to write one-off scripts
against the Shortcut API, optionally loading scdata.ari as well if that
simplifies the script.